Simplify backing up procedure
There's no need to have a remote backup script; invoke tar directly from
the local script, generating the tarball to the standard output, and capture
it from the local script directly into a local file.  This reduces the
possibility of the backup tarball being generated in one machine and we
trying to obtain it from another machine.  (It also won't use disk
space in the server for the temporary file.)